[nw] everything breaks everywhere but begin to refactor
