--- simon/agents/querymaster.py
@@ -1,101 +1,0 @@
-"""
-querymaster.py
-A few-shot classifier for knowledge queries with an llm
-"""
-
-from langchain.prompts import PromptTemplate
-from langchain.chains import LLMChain
-from langchain.output_parsers import CommaSeparatedListOutputParser
-from langchain.prompts import StringPromptTemplate
-from langchain.schema import BaseOutputParser
-
-from typing import List, Union
-from dataclasses import dataclass
-
-from pydantic import BaseModel, validator
-
-import re 
-
-from ..models import *
-
-TEMPLATE = """
-System:
-You are an AI responsible for selecting the most optimal option based on
-a query.
-
-Available options:
-{options}
-
-Selection: [*one number* representing the option you are selecting]
-
-The final line of your output must start with Selection:
-
-Begin!
-
-You are now going to select an option to {action} {input}.
-
-AI:
-Selection:"""
-
-class QuerySelectorTemplate(StringPromptTemplate):
-    options: List[QuerySelectorOption]
-    action: str = "answer the question"
-
-    def format(self, **kwargs):
-        options = "\n".join([f"Option {i}: {option.info}" for i, option in enumerate(self.options)])
-        return TEMPLATE.format(options=options, input=kwargs["input"], action=self.action)
-
-    @validator("options", each_item=True)
-    def validate_option(cls, v):
-        if type(v) == QuerySelectorOption:
-            return v
-        else:
-            raise ValueError("not a qso!")
-
-class SingleLetterOptionParser(BaseOutputParser):
-    options: List[QuerySelectorOption]
-
-    def parse(self, str):
-        try:
-            option = int(str.strip("]").strip("[").strip())
-        except ValueError:
-            results = re.search(r"option:? ?(\d+)", str, flags=re.IGNORECASE)
-            if not results:
-                breakpoint()
-            try: 
-                option = int(results.group(1).strip())
-            except ValueError:
-                option = 0
-        return self.options[option]
-
-#################
-
-class QueryMaster:
-
-    def __init__(self, context, qso, action="answer the question", verbose=False):
-        """Creates a simon query handler
-
-        Parameters
-        ----------
-        context : AgentContext
-            The context to create the assistant from.
-        qso : List[QuerySelectorOption]
-            The options for the datasources.
-        action: optional, str
-            The action that the query is selecting for
-        verbose : optional, bool
-            Whether or not the chain should be verbose.
-        """
-
-        prompt = QuerySelectorTemplate(
-            options=qso,
-            input_variables=["input"],
-            output_parser=SingleLetterOptionParser(options=qso),
-            action=action
-        )
-
-        self.__chain = LLMChain(llm=context.llm, prompt=prompt, verbose=verbose)
-
-    def __call__(self, input):
-        return self.__chain.predict_and_parse(input=input)
-
