+++ simon/agents/reason.py
@@ -8,15 +8,12 @@ from langchain.output_parsers import CommaSeparatedListOutputParser
 from langchain.prompts import BaseChatPromptTemplate
 from langchain.schema import BaseOutputParser
 
+from nltk import sent_tokenize
 from langchain.schema import (
     AIMessage,
     HumanMessage,
     SystemMessage
 )
-
-
-import logging
-L = logging.getLogger(__name__)
 
 from ..models import *
 
@@ -24,23 +21,33 @@ import re
 
 # You will be provided sources to come up with an answer. When creating your answer, adhere to the following four-element format:
 
+# Help the user come up information they wouldn't have possibly thought of regarding the Concept and fill in any gaps in knowledge they have betrayed through their question by extrapolating in a markdown list like so:
+# Provide a *full* full answer to the user's question. [4] Provide references to useful citation numbers in brackets found in the knowledge section *throughout* your answer after each claim; don't put a bunch all in the end. [2] You must only use information presented in the Knowledge section to provide your answer. [5] Use **markdown** _styles_, if appropriate. 
+
 SYSTEM_TEMPLATE = """
-Provide your response as a four-section answer. Each section should be seperated by one newline. The four sections are as follows: Citation, Answer, Reasoning, and Update. Use the following format to provide your answer. The human will provide your entire knowledeg.
+You are helping a human understand the state of a concept by being a search engine. You will be provided textual knowledge which you must refer to during your answer. At the *end* of each sentence in knowledge you are given, there is a citation take in brakets [like so] which you will refer to.
 
-```output
-Citation: 
-- "Your first quote from the knowledge. Your quote must appear exactly as-is like it appears your knowledge. Each quote may not exceed 10 words. Use ... elipses if needed."
-- "Your second quote here, it still may not exceed 10 words..."
-[this can repeat N times, but they must be quoted exactly from the Knowledge: section above. You may use elipses ... to skip sections, or [brackets] to add words to ensure flow.]
+When responding, you must provide two sections: the sections are "Answer", "Search Results". 
 
-Answer: Use the information above to come up with a markdown-formatted, easy to understand, brief answer to the user's question. The user doesn't like reading, so keep this clear and brief. Use **markdown** _styles_, if appropriate. Synthesize: don't just copy your citations.
+Answer: Provide a full, *brief (<4 sentences)*, and fact-supported answer the user's question. [2] After each of your claims, provide a tag to the sentence you used to support your claim, like so: [3]. Use **markdown** _styles_, lists, etc. if appropriate. If the Knowledge section does not provide enough information to be able to answer the question with some reasoning from you, place *ONLY* the letters N/A here. 
+Search Results: identify the sources from your search; if no results are found, place the letters N/A in this section. These should be resources from your knowledge section that directly answer the user's question, in addition to fill in any gaps of knowledge the user has betrayed through their question; respond in a markdown list:
+- *extremely* brief headline here, don't use two parts like a colon; keep it short (<10 words) [1]
+- repeat this process, provide an *very very short* headline (<10 words) and a *single* bracket link [5]
+- short headline (<10 words) and a *single* link [8]
+- ...
+- ...
+- ...
+- ...
+[This can repeat N times, but the user hates reading so keep it short. Like a search engine, put the most salient and relavent point on top, and order by relavence]
 
-Reasoning: A numbered markdown list answering three questions.
-1. Does your information contain information not in the citation section. Use 5 words to justify your answer. Begin your answer with the word "Yes", or "No".
-2. Very briefly explain if the answer is unclear, doesn't answer the prompt, or can benefit from clarification. Begin your answer with the word "Yes", or "No".
+For instance, here's an example format:
 
-Update: the characters N/A, if the first words to both of your answer to the reasoning section is "No"; otherwise, provide a better question that would answer the Human's question but would also yields outputs that satisfies both of the Reasoning conditions above
-```
+Answer: your answer here [3], some elaborations too [8]. More here.
+Search Results:
+- Brief citation headline [4]
+- Another citation headline [6]
+
+The user is smart, but is in a rush. Keep everything concise and precise.
 
 Begin!
 """
@@ -54,9 +61,7 @@ Question:
 """
 
 AI_TEMPLATE="""
-```output
-Citation:
--"""
+Answer:"""
 
 class ReasonPromptFormatter(BaseChatPromptTemplate):
     def format_messages(self, **kwargs):
@@ -69,24 +74,45 @@ class ReasonPromptFormatter(BaseChatPromptTemplate):
 class ReasonOutputParser(BaseOutputParser):
     def parse(self, str):
         str = str.strip("```output").strip("`").strip()
-        regex = r"\s*(.*)\n\n?Answer\s*:\s*(.*)\n\n?Reasoning\s*:\s*(.*)\n\n?Update\s*:\s*(.*)"
+        regex = r"\s*(.*)\n\n?Search Results\s*:\s*(.*)"
         match = re.search(regex, str, re.DOTALL)
 
-        L.debug(str)
-
         if match:
-            citation = "-"+match.group(1).strip("\"").strip('"').strip("`").strip()
-            answer = match.group(2).strip("\"").strip('"').strip("`").strip()
-            reasoning = match.group(3).strip("\"").strip('"').strip("`").strip()
-            followup = match.group(4).strip("\"").strip('"').strip("`").strip()
+            answer = match.group(1).strip("\"").strip('"').strip("`").replace("`", "").strip()
+            extrapolations = match.group(2).strip("\"").strip('"').strip("`").replace("`", "").strip()
         else:
-            return str, None, ""
+            answer = str.strip("\"").strip('"').strip("`").strip()
+            extrapolations = ""
 
+        # collect up all the [citations]
+        resource_regex = r"\[(\d+)\]"
+        resource_ids = []
+        for r in re.findall(resource_regex, str):
+            resource_ids.append(int(r.strip()))
 
-        if "n/a" in followup.lower():
-            followup = None
+        extrapolations = [i.strip()[2:].strip("\"").strip('"').strip() for i in extrapolations.split("\n")]
 
-        return answer, followup, citation
+        # collect up the extrapolation citations
+        ex_citations = []
+        for extrapolation in extrapolations:
+            # get the extrapolation id
+            results = list(re.findall(resource_regex, extrapolation))
+            if len(results) == 0:
+                ex_citations.append(-1)
+                continue
+
+            for r in results:
+                ex_citations.append(int(r.strip()))
+                continue
+        # and we now remove all the citatinos from the extrapolations
+        extrapolations = [re.sub(resource_regex, "", i).strip() for i in extrapolations]
+
+        return {
+            "answer": answer,
+            "results": [{"headline": i,
+                         "resource_id": j} for i,j in zip(extrapolations, ex_citations)],
+            "resources": resource_ids
+        }
 
 class Reason(object):
     def __init__(self, context, verbose=False):
@@ -102,9 +128,30 @@ class Reason(object):
         
         prompt = ReasonPromptFormatter(input_variables=["input", "kb"],
                                        output_parser=ReasonOutputParser())
-        self.__chain = LLMChain(llm=context.llm, prompt=prompt, verbose=verbose)
+        self.__chain = LLMChain(llm=context.reason_llm, prompt=prompt, verbose=verbose)
 
-    def __call__(self, input, kb="", entities={}):
-        return self.__chain.predict_and_parse(input=input,
-                                              kb=kb)
+    def __call__(self, input, kb="", provider="", entities={}):
+        provider_sentences = [j for i in provider.split("\n--\n") for j in sent_tokenize(i)]
+        provider_sentences = [j for i in provider_sentences for j in i.split("\n")]
+
+        num_provider_sents = len(provider_sentences)
+
+        kb_sentences = [j for i in kb.split("\n--\n") for j in sent_tokenize(i)]
+        kb_sentences = [j for i in kb_sentences for j in i.split("\n")]
+
+        sentences = kb_sentences + provider_sentences
+
+        sentence_dict = {indx:i.strip() for indx, i in enumerate(sentences)}
+        sentences = "".join([i+f" [{indx}] " for indx, i in enumerate(sentences)])
+
+        res =  self.__chain.predict_and_parse(input=input,
+                                              kb=sentences)
+        # we only leave useful resources
+        res["references"] = {int(i):sentence_dict.get(i, "") for i in res["resources"]}
+        res["context_sentence_count"] = {
+            "provider": len(provider_sentences),
+            "kb": len(kb_sentences)
+        }
+
+        return res
