--- simon/providers/__init__.py
@@ -1,5 +1,0 @@
-from .base import *
-from .kb import KnowledgeBase
-from .map import Map
-from .paper import Scholar
-
